+++ plone/recipe/codeanalysis/flake8.py
@@ -1,10 +1,9 @@
 # -*- coding: utf-8 -*-
 from plone.recipe.codeanalysis.utils import normalize_boolean
 from plone.recipe.codeanalysis.utils import read_subprocess_output
-from utils import log
-
-import os
 from tempfile import TemporaryFile
+from utils import log
+import os
 
 
 def code_analysis_flake8(options):