+++ plone/recipe/codeanalysis/tests/test_csslint.py
@@ -1,11 +1,11 @@
 # -*- coding: utf-8 -*-
-from unittest import TestCase
+from os.path import isfile as path_isfile
+from os.path import join as path_join
 from plone.recipe.codeanalysis.csslint import code_analysis_csslint
 from plone.recipe.codeanalysis.csslint import csslint_errors
 from shutil import rmtree
 from tempfile import mkdtemp
-from os.path import join as path_join
-from os.path import isfile as path_isfile
+from unittest import TestCase
 
 
 class TestCssLint(TestCase):